Initialize `bitarray_asint` as `np.int64(0)` to fix type conversion error (#180)
* Initialize bitarray_asint as np.int64

* add releasenote

* Update releasenotes/notes/fix-bitarray-asint-type-f9b12dd406e604b6.yaml

+++ qiskit_addon_sqd/qubit.py
@@ -291,7 +291,7 @@ def _int_conversion_from_bts_array(bit_array: np.ndarray) -> Any:
 
     """
     n_qubits = len(bit_array)
-    bitarray_asint = 0.0
+    bitarray_asint = np.int64(0)
     for i in range(n_qubits):
         bitarray_asint = bitarray_asint + bit_array[i] * 2 ** (n_qubits - 1 - i)
